Initial commit
The project started outside github and was not initially intended to be
public. However I though it would be a good idea if I share the project
publicly so everyone who, as me, needs an GUI app to find the duplicated
files on his/her file system to be able to benefit. So I changed the
structure and the dependencies so that it can be compiled without the
need of my internal libraries. And this is the first commit to comply
with it.