Auto configure the number of compare threads
Until now the number of compare threads was hardcoded to 4. This commit
relies on std::thread::hardware_concurrency() to get the number of
cpu cores and start that number of threads for compare. Starting twice
as much threads might be a faster option but this might lead to extreme
IO.